results numFound to long integer

+++ SolrNet/Impl/AbstractSolrQueryResults.cs
@@ -13,7 +13,7 @@ namespace SolrNet.Impl {
         /// <summary>
         /// Total documents found
         /// </summary>
-        public int NumFound { get; set; }
+        public long NumFound { get; set; }
 
         /// <summary>
         /// Start of the results
